requirements.txt
python-binance~=1.0.19
pandas~=2.1.4
python-dotenv~=1.0.0
numpy~=1.26.2
Backtesting~=0.3.3